Separate crate into library and binary halves
